#36 refactor(InquiryCommand): DTO 리팩토링
