+++ build.gradle
@@ -79,6 +79,7 @@ dependencies {
     implementation 'org.wiremock.integrations:wiremock-spring-boot:3.3.0'
 
     // ModelMapper 사용
+    //객체 간 매핑 처리
     implementation 'org.modelmapper:modelmapper:3.1.0'
 
 }
