+++ src/main/java/com/example/api/inquiry/InquiryService.java
@@ -1,7 +1,7 @@
 package com.example.api.inquiry;
 
 import com.example.api.domain.Inquiry;
-import com.example.api.inquiry.dto.RequestDTO;
+import com.example.api.inquiry.dto.InquiryRequest;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
@@ -14,7 +14,7 @@ public class InquiryService {
     private final InquiryRepository inquiryRepository;
     private final ModelMapper modelMapper;
 
-    public Inquiry saveInquiry(RequestDTO inquiryRequestDTO) {
+    public Inquiry saveInquiry(InquiryRequest inquiryRequestDTO) {
         Inquiry inquiry = modelMapper.map(inquiryRequestDTO, Inquiry.class);
         inquiry.setInquiryStatus(Inquiry.InquiryStatus.WAITING);
         inquiry.setProcessStatus("처리중");
