+++ src/main/java/com/example/api/inquiry/dto/InquiryCommand.java
@@ -1,0 +1,19 @@
+package com.example.api.inquiry.dto;
+
+import org.springframework.lang.NonNull;
+import java.time.LocalDateTime;
+
+public record InquiryCommand(
+        @NonNull
+        Long inquiryId,
+        String inquiryType,
+        String subInquiryType,
+        String title,
+        String content,
+        String inquiryStatus,
+        String processStatus,
+        LocalDateTime answerDate,
+        @NonNull
+        Long accountId
+) {
+}
