+++ src/main/java/com/example/api/inquiry/dto/InquiryResponse.java
@@ -1,0 +1,20 @@
+package com.example.api.inquiry.dto;
+
+import org.springframework.lang.NonNull;
+
+import java.time.LocalDateTime;
+
+public record InquiryResponse(
+        @NonNull
+        Long inquiryId,
+        String inquiryType,
+        String subInquiryType,
+        String title,
+        String content,
+        String inquiryStatus,
+        String processStatus,
+        LocalDateTime answerDate,
+        @NonNull
+        Long accountId
+) {
+}
